backend/requirements.txt
fastapi
uvicorn
langgraph
langchain-groq
psycopg2-binary
pydantic